+++ lib/planter/hash.rb
@@ -41,7 +41,17 @@ class ::Hash
   ## @param      second  [Hash] The hash to merge into self
   ##
   def deep_merge(second)
-    merger = proc { |_, v1, v2| Hash === v1 && Hash === v2 ? v1.merge(v2, &merger) : Array === v1 && Array === v2 ? v1 | v2 : [:undefined, nil, :nil].include?(v2) ? v1 : v2 }
+    merger = proc do |_, v1, v2|
+      if v1.is_a?(Hash) && v2.is_a?(Hash)
+        v1.merge(v2, &merger)
+      elsif v1.is_a?(Array) && v2.is_a?(Array)
+        v1 | v2
+      elsif [:undefined, nil, :nil].include?(v2)
+        v1
+      else
+        v2
+      end
+    end
     merge(second.to_h, &merger)
   end