+++ CHANGELOG.md
@@ -1,3 +1,16 @@
 ### 0.0.2
 
 2024-08-26 10:25
+
+#### NEW:
+
+- Add repo key to config, pull a git rep
+- replacements key in config for a dictionary of regex patterns and replacements
+
+#### IMPROVED
+
+- Better error reporting
+
+#### FIXED
+
+- Merge wasn't populating template placeholders